+++ src/content/index.ts
@@ -1,1 +1,3 @@
-// Instagram Detox - Content Script
+import { init as initHideExplore } from './hide-explore';
+
+initHideExplore();